refactor: change files to .tsx and add TS types
