+++ examples/with-styled-components-babel/pages/_app.tsx
