+++ examples/with-styled-components-babel/pages/index.tsx
